feat: 提供创建红线指标的服务间接口 #5598. fix
