+++ src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/resources/v3/BuildQualityIndicatorV3ResourceImpl.kt
@@ -30,32 +30,18 @@ package com.tencent.devops.quality.resources.v3
 
 import com.tencent.devops.common.api.pojo.Result
 import com.tencent.devops.common.web.RestResource
-import com.tencent.devops.quality.api.v3.ServiceQualityRuleResource
-import com.tencent.devops.quality.api.v3.pojo.request.BuildCheckParamsV3
-import com.tencent.devops.quality.api.v3.pojo.request.RuleCreateRequestV3
-import com.tencent.devops.quality.api.v3.pojo.response.RuleCreateResponseV3
-import com.tencent.devops.common.quality.pojo.RuleCheckResult
 import com.tencent.devops.quality.api.v2.pojo.request.IndicatorCreate
+import com.tencent.devops.quality.api.v3.BuildQualityIndicatorResource
 import com.tencent.devops.quality.service.v2.QualityIndicatorService
-import com.tencent.devops.quality.service.v2.QualityRuleBuildHisService
-import com.tencent.devops.quality.service.v2.QualityRuleCheckService
 import org.springframework.beans.factory.annotation.Autowired
 
 @RestResource
-class ServiceQualityRuleV3ResourceImpl @Autowired constructor(
-    private val qualityRuleBuildHisService: QualityRuleBuildHisService,
-    private val qualityRuleCheckService: QualityRuleCheckService
-) : ServiceQualityRuleResource {
-    override fun check(buildCheckParams: BuildCheckParamsV3): Result<RuleCheckResult> {
-        return Result(qualityRuleCheckService.checkBuildHis(buildCheckParams))
+class BuildQualityIndicatorV3ResourceImpl @Autowired constructor(
+    private val qualityIndicatorService: QualityIndicatorService
+) : BuildQualityIndicatorResource {
+
+    override fun upsertIndicator(userId: String, projectId: String, indicatorCreate: IndicatorCreate): Result<Boolean> {
+        return Result(qualityIndicatorService.userCreate(userId, projectId, indicatorCreate))
     }
 
-    override fun create(
-        userId: String,
-        projectId: String,
-        pipelineId: String,
-        ruleList: List<RuleCreateRequestV3>
-    ): Result<List<RuleCreateResponseV3>> {
-        return Result(qualityRuleBuildHisService.serviceCreate(userId, projectId, pipelineId, ruleList))
-    }
 }
